Suggested JS changes to vote template (#4)
This PR makes a smaller change to the primary vote listing code. Specifically,
it: uses `const` for the loop variable so that it does not pollute the global
scope, and switches to a `for..of` loop instead of a `for..in` loop, since the
latter does not guarantee iteration order (it is meant for enumerables rather
than iterables). It also switches from the `Number` coercion to the equivalent
but more concise `+` shorthand.